+++ baekjoon/bronze/b_1000.java
@@ -1,11 +1,11 @@
-package baekjoon;
+package baekjoon.bronze;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.StringTokenizer;
 
-public class b_10998 {
+public class b_1000 {
     public static void main(String[] args) throws IOException {
         // 1 2
         BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
@@ -14,6 +14,6 @@ public class b_10998 {
         int n1 = Integer.parseInt(tokenizer.nextToken());
         int n2 = Integer.parseInt(tokenizer.nextToken());
 
-        System.out.println(n1 * n2);
+        System.out.println(n1 + n2);
     }
 }
