+++ baekjoon/bronze/b_1001.java
@@ -1,19 +1,19 @@
-package baekjoon;
+package baekjoon.bronze;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.StringTokenizer;
 
-public class b_1008 {
+public class b_1001 {
     public static void main(String[] args) throws IOException {
-        // 1 3
+        // 3 2
         BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer tokenizer = new StringTokenizer(reader.readLine(), " ");
 
-        double n1 = Double.parseDouble(tokenizer.nextToken());
-        double n2 = Double.parseDouble(tokenizer.nextToken());
+        int n1 = Integer.parseInt(tokenizer.nextToken());
+        int n2 = Integer.parseInt(tokenizer.nextToken());
 
-        System.out.println(n1 / n2);
+        System.out.println(n1 - n2);
     }
 }
