+++ baekjoon/bronze/b_10430.java
@@ -1,4 +1,4 @@
-package baekjoon;
+package baekjoon.bronze;
 
 import java.io.BufferedReader;
 import java.io.IOException;
