+++ baekjoon/bronze/b_10869.java
@@ -1,19 +1,23 @@
-package baekjoon;
+package baekjoon.bronze;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.StringTokenizer;
 
-public class b_1001 {
+public class b_10869 {
     public static void main(String[] args) throws IOException {
-        // 3 2
+        // 7 3
         BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer tokenizer = new StringTokenizer(reader.readLine(), " ");
 
         int n1 = Integer.parseInt(tokenizer.nextToken());
         int n2 = Integer.parseInt(tokenizer.nextToken());
 
+        System.out.println(n1 + n2);
         System.out.println(n1 - n2);
+        System.out.println(n1 * n2);
+        System.out.println(n1 / n2);
+        System.out.println(n1 % n2);
     }
 }
