+++ baekjoon/bronze/b_18108.java
@@ -1,19 +1,17 @@
-package baekjoon;
+package baekjoon.bronze;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.util.StringTokenizer;
 
-public class b_1000 {
+public class b_18108 {
     public static void main(String[] args) throws IOException {
-        // 1 2
         BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer tokenizer = new StringTokenizer(reader.readLine(), " ");
 
-        int n1 = Integer.parseInt(tokenizer.nextToken());
-        int n2 = Integer.parseInt(tokenizer.nextToken());
+        int year = Integer.parseInt(tokenizer.nextToken());
 
-        System.out.println(n1 + n2);
+        System.out.println(year - 543);
     }
 }
