+++ baekjoon/bronze/b_2557.java
@@ -1,4 +1,4 @@
-package baekjoon;
+package baekjoon.bronze;
 
 public class b_2557 {
     public static void main(String[] args) {
